Refactor: adicionado pasta para interfaces.
