+++ DomainServices/Services/CustomerServices.cs
@@ -1,4 +1,5 @@
 ﻿using DomainModels;
+using DomainServices.Interfaces;
 using Infrastructure.Data.Context;
 using Microsoft.EntityFrameworkCore;
 using System;
